+++ lib/sidekiq/extensions/action_mailer.rb
@@ -16,7 +16,10 @@ module Sidekiq
 
       def perform(yml)
         (target, method_name, args) = YAML.load(yml)
-        target.send(method_name, *args).deliver
+        msg = target.send(method_name, *args)
+        # The email method can return nil, which causes ActionMailer to return
+        # an undeliverable empty message.
+        msg.deliver if msg && msg.to && msg.from
       end
     end